middleware/requirements.txt
paho-mqtt~=1.5.1
requests~=2.25.1
djangorestframework
markdown
django-filter
pandas
filterpy
Django~=3.1.5
psycopg2-binary